Added arm32v7 to all images
Hi,
First of all, super good initiative. Running this on K3S (Kubernetes) on Raspberry Pi's.
The setup works perfectly, it's just that the arm32v6 are not being maintained anymore.
Thus you won't get the latest version of e.g. Sonarr.
I did this minor change and it works perfectly on RPI 3 (running a cluster of 4 nodes)